+++ examples/bond/scripts/run/common/common.js
@@ -1,7 +1,8 @@
 const {
-  executeTransaction
+  executeTransaction, readAppGlobalState
 } = require('@algo-builder/algob');
 const { types } = require('@algo-builder/web');
+const { accounts } = require('./accounts');
 
 const issuePrice = 1000;
 const couponValue = 20;
@@ -120,6 +121,36 @@ function buyTx (buyer, issuerLsig, amount, algoAmount, appID, bondID) {
 }
 
 /**
+ * Returns buy transaction using algorand node
+ * @param deployer deployer object
+ * @param buyer buyer account
+ * @param issuerLsig Bond issuer logic signature
+ * @param algoAmount amount of Algo to pay for the bonds
+ * @param appID Bond application index
+ * @param bondID Bond ASA index
+ */
+async function buyTxNode (deployer, buyer, issuerLsig, algoAmount, appID, bondID) {
+  const account = await accounts(deployer);
+  const globalState = await readAppGlobalState(deployer, account.manager.addr, appID);
+  const bondPrice = globalState.get('issue_price');
+  return buyTx(buyer, issuerLsig, algoAmount / bondPrice, algoAmount, appID, bondID);
+}
+
+/**
+ * Returns buy transaction for test/runtime
+ * @param runtime runtime object
+ * @param buyer buyer account
+ * @param issuerLsig Bond issuer logic signature
+ * @param algoAmount amount of Algo to pay for the bonds
+ * @param appID Bond application index
+ * @param bondID Bond ASA index
+ */
+function buyTxRuntime (runtime, buyer, issuerLsig, algoAmount, appID, bondID) {
+  const bondPrice = Number(runtime.getGlobalState(appID, 'issue_price'));
+  return buyTx(buyer, issuerLsig, algoAmount / bondPrice, algoAmount, appID, bondID);
+}
+
+/**
  * Returns issue group transaction
  * @param creatorAccount creator acccount
  * @param issuerLsig Bond issuer logic signature
@@ -177,7 +208,8 @@ function createDexTx (
       appID: appID,
       payFlags: {},
       appArgs: ['str:create_dex'],
-      accounts: [issuerLsig.address(), dexLsig.address()]
+      accounts: [issuerLsig.address(), dexLsig.address()],
+      foreignAssets: [oldBondID]
     },
     // New bond token transfer to issuer's address
     {
@@ -291,5 +323,7 @@ module.exports = {
   buyTx,
   issueTx,
   createDexTx,
-  redeemCouponTx
+  redeemCouponTx,
+  buyTxNode,
+  buyTxRuntime
 };