+++ examples/bond/scripts/run/createDex.js
@@ -60,14 +60,7 @@ exports.createDex = async function (deployer, creatorAccount, managerAcc, i) {
   await optInTx(deployer, managerAcc, dexLsig, oldBond);
 
   const globalState = await readAppGlobalState(deployer, managerAcc.addr, appInfo.appID);
-  let total = 0;
-  for (const l of globalState) {
-    const key = Buffer.from(l.key, 'base64').toString();
-    if (key === 'total') {
-      total = l.value.uint;
-      break;
-    }
-  }
+  const total = globalState.get('total') ?? 0;
   console.log('Total issued: ', total);
 
   // balance of old bond tokens in issuer lsig